refactor: migrate to app router
