+++ packages/arb-token-bridge-ui/package.json
@@ -24,12 +24,12 @@
     "axios": "^1.7.4",
     "boring-avatars": "^1.7.0",
     "cheerio": "^1.0.0-rc.12",
-    "dayjs": "^1.11.8",
+    "dayjs": "^1.11.13",
     "ethers": "^5.6.0",
     "fetch-retry": "^5.0.3",
     "graphql": "^16.8.1",
     "lodash-es": "^4.17.21",
-    "next": "^14.2.12",
+    "next": "^14.2.15",
     "next-query-params": "^5.0.0",
     "overmind": "^28.0.1",
     "overmind-react": "^29.0.1",
