+++ packages/arb-token-bridge-ui/src/app/opengraph-image.alt
@@ -1,0 +1,1 @@
+Arbitrum Bridge
