+++ packages/arb-token-bridge-ui/src/app/page.tsx
@@ -1,0 +1,5 @@
+import Homepage from './homepage'
+
+export default async function Page() {
+  return <Homepage />
+}
