+++ packages/arb-token-bridge-ui/src/app/twitter-image.alt
@@ -1,0 +1,1 @@
+Arbitrum Bridge
