+++ packages/arb-token-bridge-ui/src/components/common/SiteBanner.tsx
@@ -1,3 +1,4 @@
+'use client'
 import { useEffect, useState } from 'react'
 import dayjs from 'dayjs'
 import { ExternalLink } from './ExternalLink'
