+++ packages/arb-token-bridge-ui/tsconfig.json
@@ -5,17 +5,32 @@
     "additional.d.ts",
     "next.config.js",
     "next-env.d.ts",
-    ".next/types/**/*.ts"
+    ".next/types/**/*.ts",
+    "build/types/**/*.ts"
   ],
   "compilerOptions": {
     "noEmit": true,
     "incremental": true,
     "jsx": "preserve",
     "paths": {
-      "@/images/*": ["./public/images/*"],
-      "@/icons/*": ["./public/icons/*"],
-      "@/token-bridge-sdk/*": ["./src/token-bridge-sdk/*"]
-    }
+      "@/images/*": [
+        "./public/images/*"
+      ],
+      "@/icons/*": [
+        "./public/icons/*"
+      ],
+      "@/token-bridge-sdk/*": [
+        "./src/token-bridge-sdk/*"
+      ]
+    },
+    "plugins": [
+      {
+        "name": "next"
+      }
+    ],
+    "strictNullChecks": true
   },
-  "exclude": ["node_modules"]
+  "exclude": [
+    "node_modules"
+  ]
 }
